feat: Move GiphyRepository to new folder structure
