+++ src/script/components/Giphy/Giphy.test.tsx
@@ -19,9 +19,8 @@
 
 import {fireEvent, render, waitFor} from '@testing-library/react';
 
+import {GiphyRepository} from 'Repositories/giphy/GiphyRepository';
 import {withTheme} from 'src/script/auth/util/test/TestUtil';
-
-import {GiphyRepository} from '../../extension/GiphyRepository';
 
 import {Giphy, GiphyState} from '.';
