+++ src/script/components/Giphy/Giphy.tsx
@@ -27,9 +27,8 @@ import {WebAppEvents} from '@wireapp/webapp-events';
 
 import {GifImage} from 'Components/Giphy/GifImage';
 import * as Icon from 'Components/Icon';
+import {Gif, GiphyRepository} from 'Repositories/giphy/GiphyRepository';
 import {t} from 'Util/LocalizerUtil';
-
-import {Gif, GiphyRepository} from '../../extension/GiphyRepository';
 
 const GIPHY_CLOSE_TIMEOUT = 350;
