+++ src/script/components/UserDevices/UserDevices.tsx
@@ -24,6 +24,7 @@ import {ClientClassification} from '@wireapp/api-client/lib/client/';
 import {useUserIdentity} from 'Hooks/useDeviceIdentities';
 import {ClientRepository, ClientEntity} from 'Repositories/client';
 import {MessageRepository} from 'Repositories/conversation/MessageRepository';
+import {CryptographyRepository} from 'Repositories/cryptography/CryptographyRepository';
 import {partition} from 'Util/ArrayUtil';
 import {t} from 'Util/LocalizerUtil';
 import {getLogger} from 'Util/Logger';
@@ -35,7 +36,6 @@ import {NoDevicesFound} from './components/NoDevicesFound';
 import {SelfFingerprint} from './components/SelfFingerprint';
 import {FIND_MODE, UserDevicesState} from './UserDevices.types';
 
-import {CryptographyRepository} from '../../cryptography/CryptographyRepository';
 import {User} from '../../entity/User';
 
 export interface UserDevicesHistoryEntry {
