+++ src/script/components/UserDevices/components/DeviceDetails/DeviceDetails.tsx
@@ -29,6 +29,7 @@ import type {ClientRepository, ClientEntity} from 'Repositories/client';
 import {isMLSConversation} from 'Repositories/conversation/ConversationSelectors';
 import {ConversationState} from 'Repositories/conversation/ConversationState';
 import type {MessageRepository} from 'Repositories/conversation/MessageRepository';
+import type {CryptographyRepository} from 'Repositories/cryptography/CryptographyRepository';
 import {WireIdentity} from 'src/script/E2EIdentity';
 import {MLSDeviceDetails} from 'src/script/page/MainContent/panels/preferences/DevicesPreferences/components/MLSDeviceDetails';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
@@ -37,7 +38,6 @@ import type {Logger} from 'Util/Logger';
 import {splitFingerprint} from 'Util/StringUtil';
 
 import {Config} from '../../../../Config';
-import type {CryptographyRepository} from '../../../../cryptography/CryptographyRepository';
 import type {User} from '../../../../entity/User';
 import {MotionDuration} from '../../../../motion/MotionDuration';
 import {FormattedId} from '../../../../page/MainContent/panels/preferences/DevicesPreferences/components/FormattedId';
