+++ src/script/components/UserDevices/components/SelfFingerprint/SelfFingerprint.tsx
@@ -26,10 +26,10 @@ import {container} from 'tsyringe';
 import {WebAppEvents} from '@wireapp/webapp-events';
 
 import {ClientState} from 'Repositories/client/ClientState';
+import type {CryptographyRepository} from 'Repositories/cryptography/CryptographyRepository';
 import {t} from 'Util/LocalizerUtil';
 import {splitFingerprint} from 'Util/StringUtil';
 
-import type {CryptographyRepository} from '../../../../cryptography/CryptographyRepository';
 import {FormattedId} from '../../../../page/MainContent/panels/preferences/DevicesPreferences/components/FormattedId';
 import {DeviceCard} from '../DeviceCard';
