+++ src/script/event/EventRepository.ts
@@ -32,6 +32,7 @@ import {Account, ConnectionState, ProcessedEventPayload} from '@wireapp/core';
 import {WebAppEvents} from '@wireapp/webapp-events';
 
 import {ClientConversationEvent, EventBuilder} from 'Repositories/conversation/EventBuilder';
+import {CryptographyMapper} from 'Repositories/cryptography/CryptographyMapper';
 import {getLogger, Logger} from 'Util/Logger';
 import {queue} from 'Util/PromiseQueue';
 import {TIME_IN_MILLIS} from 'Util/TimeUtil';
@@ -47,7 +48,6 @@ import {NOTIFICATION_HANDLING_STATE} from './NotificationHandlingState';
 import type {NotificationService} from './NotificationService';
 import {EventValidationError} from './preprocessor/EventStorageMiddleware/eventHandlers/EventValidationError';
 
-import {CryptographyMapper} from '../cryptography/CryptographyMapper';
 import {CryptographyError} from '../error/CryptographyError';
 import {EventError} from '../error/EventError';
 import type {ServerTimeHandler} from '../time/serverTimeHandler';
