+++ src/script/message/MentionEntity.ts
@@ -21,10 +21,9 @@ import {QualifiedId} from '@wireapp/api-client/lib/user';
 
 import {IMention, Mention} from '@wireapp/protocol-messaging';
 
+import {PROTO_MESSAGE_TYPE} from 'Repositories/cryptography/ProtoMessageType';
 import {matchQualifiedIds} from 'Util/QualifiedId';
 import {isUUID} from 'Util/ValidationUtil';
-
-import {PROTO_MESSAGE_TYPE} from '../cryptography/ProtoMessageType';
 
 export enum ERROR {
   INVALID_LENGTH = 'Invalid mention: Invalid length',
