+++ src/script/repositories/conversation/EventMapper.ts
@@ -25,6 +25,7 @@ import {LinkPreview, Mention} from '@wireapp/protocol-messaging';
 import {AssetRemoteData} from 'Repositories/assets/AssetRemoteData';
 import {AssetTransferState} from 'Repositories/assets/AssetTransferState';
 import {TERMINATION_REASON} from 'Repositories/calling/enum/TerminationReason';
+import {AssetData} from 'Repositories/cryptography/CryptographyMapper';
 import {t} from 'Util/LocalizerUtil';
 import {getLogger, Logger} from 'Util/Logger';
 import {userReactionMapToReactionMap} from 'Util/ReactionUtil';
@@ -43,7 +44,6 @@ import {
   CompositeMessageAddEvent,
 } from './EventBuilder';
 
-import {AssetData} from '../../cryptography/CryptographyMapper';
 import type {Conversation} from '../../entity/Conversation';
 import {Button} from '../../entity/message/Button';
 import {CallingTimeoutMessage} from '../../entity/message/CallingTimeoutMessage';
