+++ src/script/repositories/conversation/MessageRepository.test.ts
@@ -30,6 +30,7 @@ import {AudioRepository} from 'Repositories/audio/AudioRepository';
 import {ClientEntity} from 'Repositories/client/ClientEntity';
 import {ClientState} from 'Repositories/client/ClientState';
 import {ConnectionEntity} from 'Repositories/connection/ConnectionEntity';
+import {CryptographyRepository} from 'Repositories/cryptography/CryptographyRepository';
 import {Conversation} from 'src/script/entity/Conversation';
 import {Message} from 'src/script/entity/message/Message';
 import {Text} from 'src/script/entity/message/Text';
@@ -42,7 +43,6 @@ import {ConversationRepository} from './ConversationRepository';
 import {ConversationState} from './ConversationState';
 import {MessageRepository} from './MessageRepository';
 
-import {CryptographyRepository} from '../../cryptography/CryptographyRepository';
 import {ContentMessage} from '../../entity/message/ContentMessage';
 import {EventRepository} from '../../event/EventRepository';
 import {EventService} from '../../event/EventService';
