+++ src/script/repositories/cryptography/CryptographyMapper.ts
@@ -56,15 +56,15 @@ import {AssetTransferState} from 'Repositories/assets/AssetTransferState';
 import {CALL_MESSAGE_TYPE} from 'Repositories/calling/enum/CallMessageType';
 import {ConversationEphemeralHandler} from 'Repositories/conversation/ConversationEphemeralHandler';
 import {MessageAddEvent} from 'Repositories/conversation/EventBuilder';
+import {PROTO_MESSAGE_TYPE} from 'Repositories/cryptography/ProtoMessageType';
 import {getLogger, Logger} from 'Util/Logger';
 import {TIME_IN_MILLIS} from 'Util/TimeUtil';
 import {base64ToArray, arrayToBase64} from 'Util/util';
 
-import {PROTO_MESSAGE_TYPE} from '../cryptography/ProtoMessageType';
-import {CryptographyError} from '../error/CryptographyError';
-import {ClientEvent, CONVERSATION} from '../event/Client';
-import {StatusType} from '../message/StatusType';
-import {Core} from '../service/CoreSingleton';
+import {CryptographyError} from '../../error/CryptographyError';
+import {ClientEvent, CONVERSATION} from '../../event/Client';
+import {StatusType} from '../../message/StatusType';
+import {Core} from '../../service/CoreSingleton';
 
 export interface MappedText {
   data: {content: string; mentions: string[]; previews: string[]; quote?: string; replacing_message_id?: string};
