+++ src/script/repositories/giphy/GiphyRepository.ts
@@ -19,7 +19,7 @@
 
 import type {GiphyService} from './GiphyService';
 
-import {Logger, getLogger} from '../util/Logger';
+import {Logger, getLogger} from '../../util/Logger';
 
 export interface Gif {
   animated: string;
