+++ src/script/repositories/giphy/GiphyResult.ts
